move some utils code
